webapp/requirements.txt
flask
python-dotenv
requests[socks]
gunicorn
